Move `common_utils.py` into `responsibleai_dashboard` directory (#2095)
* Move `common_utils.py` into `responsibleai_dashboard` directory


* fix imports


---------
