+++ raiwidgets/tests/responsibleai_dashboard/test_responsibleai_dashboard_input_erroranalysis.py
@@ -1,11 +1,12 @@
 # Copyright (c) Microsoft Corporation
 # Licensed under the MIT License.
 import pytest
-from common_utils import CheckResponsibleAIDashboardInputTestResult
 
 from raiwidgets.interfaces import WidgetRequestResponseConstants
 from raiwidgets.responsibleai_dashboard_input import \
     ResponsibleAIDashboardInput
+
+from .common_utils import CheckResponsibleAIDashboardInputTestResult
 
 
 @pytest.mark.parametrize("with_model", [True, False])
