+++ raiwidgets/tests/responsibleai_dashboard/test_responsibleai_dashboard_input_predict.py
@@ -2,10 +2,11 @@
 # Licensed under the MIT License.
 
 import pytest
-from common_utils import CheckResponsibleAIDashboardInputTestResult
 
 from raiwidgets.responsibleai_dashboard_input import \
     ResponsibleAIDashboardInput
+
+from .common_utils import CheckResponsibleAIDashboardInputTestResult
 
 
 @pytest.mark.parametrize("with_model", [True, False])
